Add task modal added to components and containers, separated steps selectors and dispatchers on their respective containers
