+++ packages/webapp/src/components/Shift/AddTaskModal/styles.scss
@@ -1,0 +1,69 @@
+.modal {
+  width: 90%;
+  text-align: center;
+  margin: 0 auto;
+}
+.popupTitle {
+  display: flex;
+  flex-direction: row;
+  h3{
+    flex-grow: 1;
+  }
+  align-items: center;
+  margin-bottom: 10%;
+  margin-top: 2em;
+}
+
+.close {
+  width: 8%;
+  img {
+    width: 100%;
+    max-width: 40px;
+  }
+}
+
+.customContainer {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+}
+
+.taskTitle {
+  width: 100%;
+  text-align: left;
+  color: #4d4d4d;
+  font-size: 18px;
+  margin-bottom: 12px;
+  span {
+    color: var(--typeAction);
+  }
+}
+
+.taskInput {
+  width: 100%;
+  margin-bottom: 20px;
+  input {
+    width: 100%;
+    border: 2px solid #d6d2d2;
+    border-radius: 10px;
+    height: 40px;
+    margin: 0 auto;
+    padding: 0 5%;
+  }
+}
+
+.buttonContainer {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 8px;
+  margin-top: 5px;
+  padding-bottom: 0.3em;
+  button {
+    width: 90%;
+    margin: 0 auto;
+  }
+}
