+++ packages/webapp/src/containers/Shift/StepOne/StepOne.js
@@ -1,16 +1,19 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { userFarmsByFarmSelector, userFarmSelector } from "../../userFarmSlice";
-import { useTranslation } from "react-i18next";
 import PureStepOne from "../../../components/Shift/StepOne";
 import { getAllUserFarmsByFarmId } from "../../Profile/People/saga";
 import history from "../../../history";
-import { stepOneData } from "../../shiftSlice";
+import { stepOneData, stepOneSelector } from "../../shiftSlice";
+import { taskTypeSelector } from "../MyShift/selectors";
 
 function StepOne() {
   const dispatch = useDispatch();
   const users = useSelector(userFarmsByFarmSelector);
   const loggedUser = useSelector(userFarmSelector);
+  const farm = useSelector(userFarmSelector);
+  const taskTypes = useSelector(taskTypeSelector);
+  const defaultData = useSelector(stepOneSelector);
 
   useEffect(() => {
     dispatch(getAllUserFarmsByFarmId())
@@ -22,7 +25,8 @@ function StepOne() {
   }
 
   return (
-    <PureStepOne defaultWorker={{ value: loggedUser.user_id, label: loggedUser.first_name }} workers={users} onNext={onNext} onGoBack={() => history.push('/shift')} />
+    <PureStepOne defaultWorker={{ value: loggedUser.user_id, label: loggedUser.first_name }} workers={users} defaultData={defaultData}
+                 onNext={onNext} onGoBack={() => history.push('/shift')} farm={farm} taskTypes={taskTypes} />
   )
 }
