+++ packages/webapp/src/util/index.js
@@ -95,5 +95,5 @@ export const getDistance = (distance, measurement = getMeasurementFromStore()) =
 export const getDurationString = (timeInMinutes) => {
   const hours = parseInt(timeInMinutes / 60, 10);
   const minutes = timeInMinutes - (hours * 60);
-  return `${hours > 0 ? `${hours}h ` : ''}${minutes}m`
+  return `${hours > 0 ? `${hours}h ` : ''}${minutes > 0 ? `${minutes}m` : ''}`
 }
